Add haptic feedback to SegmentedRowView, SegmentRow (watchOS), and ColorPickerRowView
- Extract shared sensory feedback View extensions into HapticFeedback.swift,
  adding sensorySelectionFeedbackIfAvailable alongside the existing heavy/stop helpers
- SegmentedRowView: selection feedback on segment taps and single-mapping presses,
  heavy impact on press-release press events
- SegmentRow (watchOS): click feedback on press-release and single-mapping commits
- ColorPickerRowView: heavy impact on color circle button tap (BrightnessButton
  already had its own feedback)
